Fix rendering of rotated images
In non-square canvases, images were previously rendered incorrectly both in the getImage method and in the image selector itself in vertical rotations, due to the implicit translate caused by the swap of the x and y dimensions of the canvas.  This commit adds a translate which cancels this out and causes the images to render correctly again.